# openapi.yaml
openapi: 3.0.0
info:
  title: Moswords API
  description: Professional Team Communication Platform with Nexus Integration
  version: 1.1.0
  contact:
    name: Moswords Support
    url: https://moswords.app
  license:
    name: MIT
servers:
  - url: https://moswords.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /contacts/sync:
    post:
      summary: Batch sync multiple contacts
      tags:
        - Contacts
      operationId: batchSyncContacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - events
              properties:
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/SyncEvent'
      responses:
        '200':
          description: Contacts synced
          content:
            application/json:
              schema:
                type: object
                properties:
                  contacts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contact'
...
  /nexus/sync:
    post:
      summary: Sync contacts to Nexus Email
      tags:
        - Nexus Email
      operationId: nexusSyncContacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contacts
                - apiKey
              properties:
                contacts:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                      name:
                        type: string
                      phone:
                        type: string
                apiKey:
                  type: string
                  description: Nexus Email API key
      responses:
        '200':
          description: Contacts synced to Nexus
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  synced:
                    type: integer
                  failed:
                    type: integer
                  total:
                    type: integer
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/NexusSyncResult'

    get:
      summary: Get Nexus sync history
      tags:
        - Nexus Email
      operationId: getNexusSyncHistory
      responses:
        '200':
          description: Sync history
          content:
            application/json:
              schema:
                type: object
                properties:
                  syncHistory:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
                        contactsCount:
                          type: integer
                        successCount:
                          type: integer
                        failureCount:
                          type: integer
                        status:
                          type: string
  /nexus/os:
    post:
      summary: Interact with Nexus OS
      tags:
        - Nexus OS
      operationId: nexusOSAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
                - apiKey
              properties:
                action:
                  type: string
                  enum: [getProfile, updateProfile, getSettings, updateSettings, getConnectedApps, getActivity]
                apiKey:
                  type: string
                  description: Nexus OS API key
                profileData:
                  type: object
                settings:
                  type: object
      responses:
        '200':
          description: Nexus OS action completed
          content:
            application/json:
              schema:
                type: object

    get:
      summary: Get Nexus OS integration status
      tags:
        - Nexus OS
      operationId: getNexusOSStatus
      responses:
        '200':
          description: Integration status
          content:
            application/json:
              schema:
                type: object
                properties:
                  isConnected:
                    type: boolean
                  connectedApps:
                    type: array
                    items:
                      type: string
                  lastSync:
                    type: string
                    format: date-time
                  permissions:
                    type: array
                    items:
                      type: string
components:
  schemas:
    Contact:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        avatar:
          type: string
        userId:
          type: string
        lastSynced:
          type: string
          format: date-time
        syncedWith:
          type: array
          items:
            type: string
        customData:
          type: object
...
    SyncEvent:
      type: object
      properties:
        type:
          type: string
          enum: [add, update, delete, sync]
        contact:
          $ref: '#/components/schemas/Contact'
        timestamp:
          type: string
          format: date-time
        syncedWith:
          type: array
          items:
            type: string
    NexusSyncResult:
      type: object
      properties:
        contactId:
          type: string
        nexusId:
          type: string
        status:
          type: string
          enum: [synced, failed, pending]
        lastSync:
          type: string
          format: date-time
        error:
          type: string
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
security:
  - sessionAuth: []